enable network on ipv6-only hosts
The networkd template explicitly disables IPv6-connectivity. When
connecting to a host, the attempt to use IPv6 results in -ENETUNREACH
from the guest kernel. If the host is IPv6-only, the host kernel
likewise makes attempts to use IPv4 result in -ENETUNREACH. Hence
fakemachine has dysfunctional network when invoked on an IPv6-only host.

Closes: #207
Signed-off-by: Helmut Grohne <helmut@subdivi.de>